Renamed the Key variant class files to match new class names
